Refactor to split event processing from the apply_event headers
