+++ lib/lattice_observer/observed/lattice.ex
@@ -7,7 +7,16 @@ defmodule LatticeObserver.Observed.Lattice do
   lattice events
   """
   alias __MODULE__
-  alias LatticeObserver.Observed.{Provider, Host, Actor, Instance, LinkDefinition, Decay}
+
+  alias LatticeObserver.Observed.{
+    Provider,
+    Host,
+    Actor,
+    Instance,
+    LinkDefinition,
+    Decay,
+    EventProcessor
+  }
 
   require Logger
 
@@ -54,7 +63,7 @@ defmodule LatticeObserver.Observed.Lattice do
           parameters: [Parameters.t()]
         }
 
-  @spec new(keyword) :: t()
+  @spec new(Keyword.t()) :: LatticeObserver.Observed.Lattice.t()
   def new(parameters \\ []) do
     %Lattice{
       actors: %{},
@@ -90,7 +99,7 @@ defmodule LatticeObserver.Observed.Lattice do
           type: "com.wasmcloud.lattice.host_heartbeat"
         }
       ) do
-    record_heartbeat(l, source_host, stamp, data)
+    EventProcessor.record_heartbeat(l, source_host, stamp, data)
   end
 
   def apply_event(
@@ -104,7 +113,7 @@ defmodule LatticeObserver.Observed.Lattice do
         }
       ) do
     labels = Map.get(data, "labels", %{})
-    record_host(l, source_host, labels, stamp)
+    EventProcessor.record_host(l, source_host, labels, stamp)
   end
 
   def apply_event(
@@ -115,7 +124,7 @@ defmodule LatticeObserver.Observed.Lattice do
           type: "com.wasmcloud.lattice.host_stopped"
         }
       ) do
-    remove_host(l, source_host)
+    EventProcessor.remove_host(l, source_host)
   end
 
   def apply_event(
@@ -168,7 +177,7 @@ defmodule LatticeObserver.Observed.Lattice do
     spec = Map.get(annotations, @annotation_app_spec, "")
     claims = Map.get(d, "claims", %{})
 
-    put_provider_instance(
+    EventProcessor.put_provider_instance(
       l,
       source_host,
       pk,
@@ -211,7 +220,7 @@ defmodule LatticeObserver.Observed.Lattice do
       ) do
     annotations = Map.get(d, "annotations", %{})
     spec = Map.get(annotations, @annotation_app_spec, "")
-    remove_provider_instance(l, source_host, pk, link_name, instance_id, spec)
+    EventProcessor.remove_provider_instance(l, source_host, pk, link_name, instance_id, spec)
   end
 
   def apply_event(
@@ -228,7 +237,7 @@ defmodule LatticeObserver.Observed.Lattice do
         }
       ) do
     spec = Map.get(d, "annotations", %{}) |> Map.get(@annotation_app_spec, "")
-    remove_actor_instance(l, source_host, pk, instance_id, spec)
+    EventProcessor.remove_actor_instance(l, source_host, pk, instance_id, spec)
   end
 
   def apply_event(
@@ -247,7 +256,7 @@ defmodule LatticeObserver.Observed.Lattice do
       ) do
     spec = Map.get(d, "annotations", %{}) |> Map.get(@annotation_app_spec, "")
     claims = Map.get(d, "claims", %{})
-    put_actor_instance(l, source_host, pk, instance_id, spec, stamp, claims)
+    EventProcessor.put_actor_instance(l, source_host, pk, instance_id, spec, stamp, claims)
   end
 
   def apply_event(
@@ -278,7 +287,7 @@ defmodule LatticeObserver.Observed.Lattice do
           type: "com.wasmcloud.lattice.linkdef_set"
         }
       ) do
-    put_linkdef(l, actor_id, link_name, provider_id, contract_id, values)
+    EventProcessor.put_linkdef(l, actor_id, link_name, provider_id, contract_id, values)
   end
 
   def apply_event(
@@ -295,7 +304,7 @@ defmodule LatticeObserver.Observed.Lattice do
           type: "com.wasmcloud.lattice.linkdef_deleted"
         }
       ) do
-    del_linkdef(l, actor_id, link_name, provider_id)
+    EventProcessor.del_linkdef(l, actor_id, link_name, provider_id)
   end
 
   def apply_event(l = %Lattice{}, %Cloudevents.Format.V_1_0.Event{
@@ -327,267 +336,13 @@ defmodule LatticeObserver.Observed.Lattice do
         type: "com.wasmcloud.synthetic.decay_ticked",
         time: stamp
       }) do
-    event_time = timestamp_from_iso8601(stamp)
+    event_time = EventProcessor.timestamp_from_iso8601(stamp)
     Decay.age_hosts(l, event_time)
   end
 
   def apply_event(l = %Lattice{}, evt) do
     Logger.warn("Unexpected event: #{inspect(evt)}")
     l
-  end
-
-  defp put_linkdef(l = %Lattice{}, actor_id, link_name, provider_id, contract_id, values) do
-    case Enum.find(l.linkdefs, fn link ->
-           link.actor_id == actor_id && link.provider_id == provider_id &&
-             link.link_name == link_name
-         end) do
-      nil ->
-        ld = %LinkDefinition{
-          actor_id: actor_id,
-          link_name: link_name,
-          provider_id: provider_id,
-          contract_id: contract_id,
-          values: values
-        }
-
-        %Lattice{l | linkdefs: [ld | l.linkdefs]}
-
-      _ ->
-        l
-    end
-  end
-
-  defp del_linkdef(l = %Lattice{}, actor_id, link_name, provider_id) do
-    %Lattice{
-      l
-      | linkdefs:
-          Enum.reject(l.linkdefs, fn link ->
-            link.actor_id == actor_id && link.link_name == link_name &&
-              link.provider_id == provider_id
-          end)
-    }
-  end
-
-  defp remove_actor_instance(l = %Lattice{}, _host_id, pk, instance_id, _spec) do
-    actor = l.actors[pk]
-
-    if actor != nil do
-      actor = %Actor{
-        actor
-        | instances: actor.instances |> Enum.reject(fn i -> i.id == instance_id end)
-      }
-
-      %Lattice{
-        l
-        | actors: Map.put(l.actors, pk, actor),
-          instance_tracking: l.instance_tracking |> Map.delete(instance_id)
-      }
-      |> strip_instanceless_entities()
-    else
-      l
-    end
-  end
-
-  defp put_actor_instance(l = %Lattice{}, host_id, pk, instance_id, spec, stamp, claims)
-       when is_binary(pk) and is_binary(instance_id) and is_binary(spec) do
-    actor =
-      Map.get(l.actors, pk, %Actor{
-        id: pk,
-        name: Map.get(claims, "name", "unavailable"),
-        capabilities: Map.get(claims, "caps", []),
-        issuer: Map.get(claims, "issuer", ""),
-        tags: Map.get(claims, "tags", []),
-        call_alias: Map.get(claims, "call_alias", ""),
-        instances: []
-      })
-
-    instance = %Instance{
-      id: instance_id,
-      host_id: host_id,
-      spec_id: spec,
-      version: Map.get(claims, "version", ""),
-      revision: Map.get(claims, "revision", 0)
-    }
-
-    actor =
-      if actor.instances |> Enum.find(fn i -> i.id == instance_id end) == nil do
-        %{actor | instances: [instance | actor.instances]}
-      else
-        actor
-      end
-
-    %Lattice{
-      l
-      | actors: Map.put(l.actors, pk, actor),
-        instance_tracking:
-          Map.put(l.instance_tracking, instance.id, timestamp_from_iso8601(stamp))
-    }
-  end
-
-  defp put_provider_instance(
-         l = %Lattice{},
-         source_host,
-         pk,
-         link_name,
-         contract_id,
-         instance_id,
-         spec,
-         stamp,
-         claims
-       ) do
-    provider =
-      Map.get(l.providers, {pk, link_name}, %Provider{
-        id: pk,
-        name: Map.get(claims, "name", "unavailable"),
-        issuer: Map.get(claims, "issuer", ""),
-        contract_id: contract_id,
-        tags: Map.get(claims, "tags", []),
-        link_name: link_name,
-        instances: []
-      })
-
-    instance = %Instance{
-      id: instance_id,
-      host_id: source_host,
-      spec_id: spec,
-      version: Map.get(claims, "version", ""),
-      revision: Map.get(claims, "revision", 0)
-    }
-
-    provider =
-      if provider.instances |> Enum.find(fn i -> i.id == instance_id end) == nil do
-        %{provider | instances: [instance | provider.instances]}
-      else
-        provider
-      end
-
-    %Lattice{
-      l
-      | providers: Map.put(l.providers, {pk, link_name}, provider),
-        instance_tracking:
-          Map.put(l.instance_tracking, instance.id, timestamp_from_iso8601(stamp))
-    }
-  end
-
-  defp remove_provider_instance(l, _source_host, pk, link_name, instance_id, _spec) do
-    provider = l.providers[{pk, link_name}]
-
-    if provider != nil do
-      provider = %Provider{
-        provider
-        | instances: provider.instances |> Enum.reject(fn i -> i.id == instance_id end)
-      }
-
-      %Lattice{
-        l
-        | providers: Map.put(l.providers, {pk, link_name}, provider),
-          instance_tracking: l.instance_tracking |> Map.delete(instance_id)
-      }
-      |> strip_instanceless_entities()
-    else
-      l
-    end
-  end
-
-  defp remove_host(l = %Lattice{}, source_host) do
-    # NOTE: the instance_tracking map will purge unseen instances
-    # during the decay event processing.
-
-    l = %Lattice{
-      l
-      | hosts: Map.delete(l.hosts, source_host),
-        actors:
-          l.actors
-          |> Enum.map(fn {k, v} ->
-            {k,
-             %Actor{
-               v
-               | instances: v.instances |> Enum.reject(fn i -> i.host_id == source_host end)
-             }}
-          end)
-          |> Enum.into(%{}),
-        providers:
-          l.providers
-          |> Enum.map(fn {k, v} ->
-            {k,
-             %Provider{
-               v
-               | instances: v.instances |> Enum.reject(fn i -> i.host_id == source_host end)
-             }}
-          end)
-          |> Enum.into(%{})
-    }
-
-    l |> strip_instanceless_entities()
-  end
-
-  defp strip_instanceless_entities(l = %Lattice{}) do
-    %Lattice{
-      l
-      | actors: l.actors |> Enum.reject(fn {_k, v} -> v.instances == [] end) |> Enum.into(%{}),
-        providers:
-          l.providers |> Enum.reject(fn {_k, v} -> v.instances == [] end) |> Enum.into(%{})
-    }
-  end
-
-  defp record_heartbeat(l = %Lattice{}, source_host, stamp, data) do
-    labels = Map.get(data, "labels", %{})
-    l = record_host(l, source_host, labels, stamp)
-
-    l =
-      List.foldl(Map.get(data, "actors", []), l, fn x, acc ->
-        # TODO - once the host can emit annotations we can pull the spec
-        put_actor_instance(acc, source_host, x["public_key"], x["instance_id"], "", stamp, %{})
-      end)
-
-    l =
-      List.foldl(Map.get(data, "providers", []), l, fn x, acc ->
-        # TODO - once the host emits annotations we can pull the spec
-        put_provider_instance(
-          acc,
-          source_host,
-          x["public_key"],
-          x["link_name"],
-          x["contract_id"],
-          x["instance_id"],
-          "",
-          stamp,
-          %{}
-        )
-      end)
-
-    l
-  end
-
-  defp record_host(l = %Lattice{}, source_host, labels, stamp) do
-    host =
-      Map.get(l.hosts, source_host, %Host{
-        id: source_host,
-        labels: labels,
-        first_seen: timestamp_from_iso8601(stamp)
-      })
-
-    # Every time we see a host, we set the last seen stamp
-    # and bump it to healthy (TODO: support aggregate status based on
-    # host contents)
-    host =
-      Map.merge(
-        host,
-        %{
-          last_seen: timestamp_from_iso8601(stamp),
-          status: :healthy
-        }
-      )
-
-    %Lattice{l | hosts: Map.put(l.hosts, source_host, host)}
-  end
-
-  @spec timestamp_from_iso8601(binary) :: DateTime.t()
-  def timestamp_from_iso8601(stamp) when is_binary(stamp) do
-    case DateTime.from_iso8601(stamp) do
-      {:ok, datetime, 0} -> datetime
-      _ -> DateTime.utc_now()
-    end
   end
 
   @spec running_instances(
